requirements.txt
Flask==2.3.3
opencv-python==4.9.0.80
numpy==2.2.4
Pillow==10.3.0
pandas==2.2.2
face-recognition==1.3.0
dlib==19.24.2
Werkzeug==2.3.8
